Rename internal packages, avoid package name shared across artifacts
Just package renames - no functional changes.

(Re)using the same package name in different artifacts is not just only
bad practice, it can likely cause issues later down the road, e.g. with
JPMS, but also Quarkus warns about that.
